requirements.txt
-e git://github.com/magnet-cl/tablib.git@master#egg=tablib
Django==1.8.18
Fabric==1.10.2
Faker==0.8.3
bootstrap-admin==0.3.6
django-bootstrap-form==3.2.1
django-ckeditor==5.2.1
django-compressor==2.0
django-countries==4.4
django-cron==0.5.0
django-debug-toolbar==1.7
django-filter==1.0.1
django-form-utils==1.0.3
django-formtools==1.0
django-import-export==0.5.1
django-loginas==0.3.2
django-pdfkit==0.2.1
django-phonenumber-field==1.1.0
django-recaptcha==1.0.5
django-robots==3.0
django-ses==0.8.2
djangorestframework==3.5.3
docutils==0.13.1
easy-thumbnails==2.3
filemagic==1.6
ipdb==0.10.3
ipython==4.0.0
jsonfield==1.0.3
markdown==2.6.7
model-mommy==1.3.2
pdfkit==0.6.0
psycopg2==2.6.1
pycrypto==2.6.1
pyexcel-xlsx==0.2.3
pyjade==3.1.0
python-dateutil==2.6.0
pytz>=2016.4
requests==2.13.0
